Create updateGameList.sh
Hi, this script is not connected to nothing, I've used it to generate the gamelist based on linuxgsm.sh list cache.
I've sent it to jesinmat hoping he would attach it to the updateGamelList instead of pulling from his repo. Its been several months and I have no answers. 
The script generates the folder structure and minimum requirements for the games to be added to the list 

I hope it can be of any use since iv'e seen that jesinmat also stopped commiting pull requests for his game repo, which is attached to this appliance. 

The script is able to keep the gamelist up to date with the one from linuxGSM